IT WORKS.
And by that I mean something along the lines of "there is a block, and
it is capable of converting water into a crappy-ass implementation of
steam when supplied with some form of fuel (i.e. lava), as well as
interacting with fluid containers and pipes."

What more could you ask for? Apart from a GUI, mod integration, active
steam output, support for other types of fuel, the ability to run on
items, animations, particles, server compatibility...